matlab: improve data loading and writing
some of the manipulations matlab requires for converting from bytes to the correct data format are non obvious, so ornot classes perform some more that they previously did.

Additionally some of the existing functionality was broken for complex data and half data.

+++ matlab/+ornot/beamformSimpleParameters.m
@@ -11,13 +11,13 @@ end
 
 switch bsp.data_kind
     case OGLBeamformerDataKind.Int16
-        data_size_multiplier = 2;
+        data_size_multiplier = uint32(2);
     case OGLBeamformerDataKind.Int16Complex
-        data_size_multiplier = 4;
+        data_size_multiplier = uint32(4);
     case OGLBeamformerDataKind.Float32
-        data_size_multiplier = 4;
+        data_size_multiplier = uint32(4);
     case OGLBeamformerDataKind.Float32Complex
-        data_size_multiplier = 8;
+        data_size_multiplier = uint32(8);
 end
 data_size = data_size_multiplier*prod(bsp.raw_data_dimensions);
